making the requested changes
