+++ src/mocks/course.ts
@@ -1,0 +1,31 @@
+import theme from '../styles/theme';
+
+const COURSE_DATA = {
+  frontEnd: {
+    backgroundTitle: '#0d6f61',
+    widthContainerTitle: '300px',
+    title: 'Programação Front-end',
+    coursePresentation: '',
+    courseModules: ['Como funciona a internet', 'HTML5', 'CSS3'],
+  },
+  backEnd: {
+    backgroundTitle: '#1e6f7a',
+    widthContainerTitle: '293px',
+    title: 'Programação Back-end',
+    coursePresentation: '',
+    courseModules: ['Lógica de programação', 'Arquitetura de dados', 'nodeJS'],
+  },
+  design: {
+    backgroundTitle: '#45408e',
+    widthContainerTitle: '173px',
+    title: 'Design UI/UX',
+    coursePresentation: '',
+    courseModules: [
+      'Figma',
+      'O que é experiência do usuário',
+      'Design de interface',
+    ],
+  },
+};
+
+export { COURSE_DATA };
